provide guest bindings if compiled for wasm32 target
